Move Advanced Settings Menu to its own file
